Fix quality issue rendering on high zoom levels (#176)

+++ src/main/java/com/github/thecoldwine/sigrun/common/ext/GoogleCoordUtils.java
@@ -87,6 +87,6 @@ public class GoogleCoordUtils {
 
 		double earthCircumference = 2 * Math.PI * EARTH_R;
 		return Math.cos(Math.toRadians(latlon.getLatDgr())) * earthCircumference
-				/ (TILE_SIZE * scale);
+				/ ((long)TILE_SIZE * scale);
 	}
 }
